extract NamesLookup to own file (jdk-bug enum inner classes)
